footerコンポーネントの切り分け
