+++ src/app/layout.module.scss
@@ -41,7 +41,6 @@
   }
 }
 
-// PC footerのスタイル
 .site-title {
   display: flex;
   gap: 10px;
@@ -53,31 +52,4 @@
   @include hover-active {
     opacity: 0.7;
   }
-}
-
-.site-list {
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-  font-size: 24px;
-  line-height: 29px;
-}
-
-%site-link {
-  display: flex;
-  gap: 12px;
-  align-items: center;
-
-  @include hover-active {
-    opacity: 0.7;
-  }
-}
-
-.site-link {
-  @extend %site-link;
-}
-
-.site-link-active {
-  @extend %site-link;
-  color: #e54d2e;
 }
