+++ src/app/layout.tsx
@@ -1,11 +1,10 @@
-"use client"; // TODO: usePathnameを利用するために設定しているが、該当箇所以外はServer側で生成されるように調整したい
-
 import "@/styles/globals.scss";
 
 import { Inter } from "next/font/google";
 import Image from "next/image";
 import Link from "next/link";
-import { usePathname } from "next/navigation";
+
+import { Footer } from "@/components/Footer";
 
 import styles from "./layout.module.scss";
 
@@ -16,8 +15,6 @@ export default function RootLayout({
 }: {
   children: React.ReactNode;
 }) {
-  const pagePath = usePathname();
-
   return (
     <html lang="ja">
       <body className={inter.className}>
@@ -30,75 +27,7 @@ export default function RootLayout({
                   Top Chefs
                 </Link>
               </h1>
-              <ul className={styles["site-list"]}>
-                <li>
-                  <Link
-                    href="/"
-                    className={
-                      pagePath === "/"
-                        ? styles["site-link-active"]
-                        : styles["site-link"]
-                    }
-                  >
-                    <Image
-                      src={
-                        // TODO: SVGをReactコンポーネントとして利用してpropsで色変更できるようにしたい
-                        pagePath === "/"
-                          ? "/icon/search-active.svg"
-                          : "/icon/search.svg"
-                      }
-                      width={26}
-                      height={26}
-                      alt=""
-                    />
-                    話題を検索
-                  </Link>
-                </li>
-                <li>
-                  <Link
-                    href="/favorites"
-                    className={
-                      pagePath === "/favorites"
-                        ? styles["site-link-active"]
-                        : styles["site-link"]
-                    }
-                  >
-                    <Image
-                      src={
-                        pagePath === "/favorites"
-                          ? "/icon/heart-active.svg"
-                          : "/icon/heart.svg"
-                      }
-                      width={26}
-                      height={26}
-                      alt=""
-                    />
-                    お気に入り
-                  </Link>
-                </li>
-                <li>
-                  <Link
-                    href="/shopping-lists"
-                    className={
-                      pagePath === "/shopping-lists"
-                        ? styles["site-link-active"]
-                        : styles["site-link"]
-                    }
-                  >
-                    <Image
-                      src={
-                        pagePath === "/shopping-lists"
-                          ? "/icon/cart-active.svg"
-                          : "/icon/cart.svg"
-                      }
-                      width={26}
-                      height={26}
-                      alt=""
-                    />
-                    買い物リスト
-                  </Link>
-                </li>
-              </ul>
+              <Footer />
             </div>
             <div className={styles["layout-page"]}>{children}</div>
           </div>
