+++ src/components/Footer/Footer.module.scss
@@ -1,0 +1,28 @@
+@import "@/styles/mixin";
+
+.site-list {
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+  font-size: 24px;
+  line-height: 29px;
+}
+
+%site-link {
+  display: flex;
+  gap: 12px;
+  align-items: center;
+
+  @include hover-active {
+    opacity: 0.7;
+  }
+}
+
+.site-link {
+  @extend %site-link;
+}
+
+.site-link-active {
+  @extend %site-link;
+  color: #e54d2e;
+}
