+++ src/components/Footer/Footer.stories.tsx
@@ -1,0 +1,9 @@
+import { Footer } from "@/components/Footer";
+
+export default {
+  component: Footer,
+};
+
+export const Default = {
+  args: {},
+};
